+++ server/party_handler_test.go
@@ -61,9 +61,10 @@ func createTestPartyHandler(t *testing.T, logger *zap.Logger) (*PartyHandler, fu
 	mm, cleanup, _ := createTestMatchmaker(t, logger, true, nil)
 	tt := testTracker{}
 	tsm := testStreamManager{}
+
 	dmr := DummyMessageRouter{}
 
-	pr := NewLocalPartyRegistry(logger, mm, &tt, &tsm, &dmr, node)
+	pr := NewLocalPartyRegistry(logger, cfg, mm, &tt, &tsm, &dmr, node)
 	ph := NewPartyHandler(logger, pr, mm, &tt, &tsm, &dmr, uuid.UUID{}, node, true, 10, nil)
 	return ph, cleanup
 }